Fix for PbtA bundling
