+++ module/helpers/traits-sheet.mjs
@@ -1,5 +1,3 @@
-import PbtaItemSheet from "../../../../systems/pbta/module/applications/item/item-sheet.js";
-
 // Define data model for trait sheet
 export class RootTraitsModel extends foundry.abstract.TypeDataModel {
     static defineSchema() {
@@ -13,23 +11,5 @@ export class RootTraitsModel extends foundry.abstract.TypeDataModel {
 
     prepareDerivedData() {
       this.nSteps = this.steps.length;
-    }
-}
-
-// Extend PbtA item sheets and change template path
-export class RootTraitsSheet extends PbtaItemSheet {
-
-    get template() {
-        return `/modules/root/templates/traits-sheet.hbs`;
-    }
-
-    async getData(options = {}) {
-        const context = await super.getData(options);
-        context.description = await TextEditor.enrichHTML(this.object.system.description, {
-            async: true,
-            secrets: this.object.isOwner,
-            relativeTo: this.object
-        });
-        return context;
     }
 }
